rename global_metrics package to metrics
