+++ internal/metrics/metrics.go
@@ -1,4 +1,4 @@
-package global_metrics
+package metrics
 
 import (
 	"sync"
